@@ -8,8 +8,12 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "@nutui/nutui": "^4.0.0",
+    "ansi-regex": "^6.0.1",
     "core-js": "^3.6.5",
-    "vue": "^2.6.11"
+    "vant": "^4.0.10",
+    "vue": "^3.2.47",
+    "vue-router": "^4.1.6"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
